+++ R/utils.R
@@ -18,10 +18,21 @@ create_download_dir <- function(pattern)
   ))
 }
 
-# dav_path ---------------------------------------------------------------------
-dav_path <- function()
+# webdav_base_url --------------------------------------------------------------
+webdav_base_url <- function(leading_slash = TRUE)
 {
-  "/remote.php/dav"
+  # The base url for all WebDAV operations for a Nextcloud instance is
+  # /remote.php/dav.
+  # https://docs.nextcloud.com/server/latest/developer_manual/client_apis/WebDAV/basic.html
+
+  path <- "remote.php/dav"
+
+
+  if (!leading_slash) {
+    return(path)
+  }
+
+  paste0("/", path)
 }
 
 # decode_url -------------------------------------------------------------------
@@ -42,7 +53,7 @@ exclude_directories <- function(file_info)
 # fileid_to_version_href -------------------------------------------------------
 fileid_to_version_href <- function(fileid = "", user = nextcloud_user())
 {
-  file.path(dav_path(), "versions", user, "versions", fileid)
+  file.path(webdav_base_url(), "versions", user, "versions", fileid)
 }
 
 # get_file_or_folder_info ------------------------------------------------------
@@ -60,7 +71,7 @@ get_file_or_folder_info <- function(
 # href_to_url ------------------------------------------------------------------
 href_to_url <- function(href)
 {
-  file.path(nextcloud_url(), href)
+  file.path(nextcloud_url(), remove_leading_slashes(href))
 }
 
 # indent -----------------------------------------------------------------------
@@ -128,9 +139,11 @@ nextcloud_auth <- function(
 }
 
 # path_to_file_href ------------------------------------------------------------
-path_to_file_href <- function(path = "", user = nextcloud_user())
+path_to_file_href <- function(
+  path = "", user = nextcloud_user(), leading_slash = FALSE
+)
 {
-  file.path(dav_path(), "files", user, path)
+  file.path(webdav_base_url(leading_slash = leading_slash), "files", user, path)
 }
 
 # remove_leading_slashes -------------------------------------------------------