+++ tests/testthat/test-function-path_to_file_href.R
@@ -1,7 +1,7 @@
 test_that("path_to_file_href() works", {
 
-  expect_true(
-    startsWith(kwb.nextcloud:::path_to_file_href(), "/remote.php/dav/files")
-  )
+  f <- kwb.nextcloud:::path_to_file_href
 
+  expect_true(startsWith(f(leading_slash = TRUE), "/remote.php/dav/files"))
+  expect_true(startsWith(f(leading_slash = FALSE), "remote.php/dav/files"))
 })